requirements.txt
# General packages
numpy==1.20.1
pandas==1.2.2
scikit_learn==0.24.1
tqdm==4.57.0
PyYAML==5.4.1
jupyter==1.0.0
matplotlib==3.3.1

# DVC
dvc[s3]==2.1.0

# PyTorch Geometric
--find-links https://download.pytorch.org/whl/torch_stable.html
--find-links https://pytorch-geometric.com/whl/torch-1.7.0+cu110.html

torch==1.7.0+cu110
torch-scatter==2.0.5
torch-sparse==0.6.8
torch-cluster==1.5.8
torch-spline-conv==1.2.0
torch-geometric==1.7.0

# Open Graph Benchmark
ogb>=1.3.0

# Lightning-bolts for LR scheduler
lightning-bolts==0.3.3
